runtime: inline ts format owner module
