--- crates/ts-format-core/README.md
@@ -1,10 +1,0 @@
-# adze-ts-format-core
-
-Tree-sitter action-tag constants and action-priority selection helpers.
-
-Part of the [adze](https://github.com/EffortlessMetrics/adze) workspace.
-
-## License
-
-Licensed under either of [Apache License, Version 2.0](../../LICENSE-APACHE)
-or [MIT License](../../LICENSE-MIT) at your option.
