+++ runtime/Cargo.toml
@@ -68,7 +68,6 @@ smallvec.workspace = true
 wasm-bindgen = { version = "0.2.100", optional = true }
 adze-concurrency-caps-core = { version = "0.1.0", path = "../crates/concurrency-caps-core" }
 adze-linecol-core = { version = "0.1.0", path = "../crates/linecol-core" }
-adze-ts-format-core = { version = "0.1.0", path = "../crates/ts-format-core" }
 ts_bridge = { package = "ts-bridge", version = "0.1.0", path = "../tools/ts-bridge", optional = true, default-features = false }
 tree_sitter_json = { package = "tree-sitter-json", version = "0.24", optional = true }
 adze-runtime-governance-api = { version = "0.1.0", path = "../crates/runtime-governance-api" }
