+++ runtime/tests/ts_format_microcrate_integration.rs
@@ -1,15 +1,9 @@
 use adze::ts_format::{TSActionTag, choose_action, choose_action_with_precedence};
 use adze_glr_core::{Action, ParseTable, RuleId, StateId};
-use adze_ts_format_core as core_ts_format;
 
 #[test]
-fn runtime_reexport_matches_microcrate_behavior() {
-    let mut runtime_table = ParseTable {
-        dynamic_prec_by_rule: vec![2, 5],
-        rule_assoc_by_rule: vec![-1, 1],
-        ..ParseTable::default()
-    };
-    let mut core_table = ParseTable {
+fn runtime_ts_format_selects_stable_actions() {
+    let mut parse_table = ParseTable {
         dynamic_prec_by_rule: vec![2, 5],
         rule_assoc_by_rule: vec![-1, 1],
         ..ParseTable::default()
@@ -22,23 +16,24 @@ fn runtime_reexport_matches_microcrate_behavior() {
         Action::Error,
     ];
 
-    assert_eq!(choose_action(&cell), core_ts_format::choose_action(&cell));
+    assert_eq!(choose_action(&cell), Some(Action::Accept));
     assert_eq!(
-        choose_action_with_precedence(&cell, &runtime_table),
-        core_ts_format::choose_action_with_precedence(&cell, &core_table)
+        choose_action_with_precedence(&cell, &parse_table),
+        Some(Action::Accept)
     );
 
-    runtime_table.dynamic_prec_by_rule[0] = 1;
-    core_table.dynamic_prec_by_rule[0] = 1;
+    parse_table.dynamic_prec_by_rule[0] = 1;
     assert_eq!(
-        choose_action_with_precedence(&cell, &runtime_table),
-        core_ts_format::choose_action_with_precedence(&cell, &core_table)
+        choose_action_with_precedence(&cell, &parse_table),
+        Some(Action::Accept)
     );
 }
 
 #[test]
-fn runtime_reexport_stays_type_compatible() {
-    let runtime_tag = TSActionTag::Accept as u8;
-    let core_tag = core_ts_format::TSActionTag::Accept as u8;
-    assert_eq!(runtime_tag, core_tag);
+fn runtime_ts_format_tag_values_match_tree_sitter() {
+    assert_eq!(TSActionTag::Error as u8, 0);
+    assert_eq!(TSActionTag::Shift as u8, 1);
+    assert_eq!(TSActionTag::Recover as u8, 2);
+    assert_eq!(TSActionTag::Reduce as u8, 3);
+    assert_eq!(TSActionTag::Accept as u8, 4);
 }
